gpu: msm: a5xx: fix busy data usage reports
The driver reports the GPU usage percentage to the userspace via sysfs.
The calculation of this value is managed by the 'a5xx_read_throttling_counters'
function.

In particular, the 'counter_delta' function is called to populate the 'th'
array.
Here the issue:
- 'counter_delta' returns an unsigned int,
- the 'th' array contains uint32_t values,
- the 'adj' variable, used in the calculation, is a 'int64_t'.

To get the correct results, use 'th' and 'adj' as uint64_t.

+++ drivers/gpu/msm/adreno_a5xx.c
@@ -1697,8 +1697,8 @@ static int a5xx_enable_pwr_counters(struct adreno_device *adreno_dev,
 static int64_t a5xx_read_throttling_counters(struct adreno_device *adreno_dev)
 {
 	int i;
-	int64_t adj;
-	uint32_t th[ADRENO_GPMU_THROTTLE_COUNTERS];
+	uint64_t adj;
+	uint64_t th[ADRENO_GPMU_THROTTLE_COUNTERS];
 	struct adreno_busy_data *busy = &adreno_dev->busy_data;
 
 	if (!adreno_is_a540(adreno_dev))
